Formatted and moved CreationPreviewNavigationTest in a separate file
